Admin Dashboard DONEEE front and back edit and delete

+++ routes/admin.route.js
@@ -11,7 +11,7 @@ router.get("/", isLoggedIn, async (req, res) => {
   } else {
     try {
       let users = await User.find({});
-      let projects = await Project.find({});
+      let projects = await Project.find({}).populate("user");
       res.status(200).json({ users, projects });
     } catch (error) {
       res.json(error);
@@ -38,26 +38,51 @@ router.get("/delete/user/:id", isLoggedIn, async (req, res) => {
 });
 
 router.get("/delete/project/:id", async (req, res) => {
-  try {
-    let project = await Project.findByIdAndDelete(req.params.id);
-    if (!project) throw { message: "Couldn't find the project" };
-    res.status(200).json({ project });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ message: "something went south check the logs" });
+  if (!req.user.isSuperAdmin) {
+    res.status(403).json("Not Cool");
+  } else {
+    try {
+      let project = await Project.findByIdAndDelete(req.params.id);
+      if (!project) throw { message: "Couldn't find the project" };
+      res.status(200).json({ project });
+    } catch (error) {
+      console.log(error);
+      res.status(500).json({ message: "something went south check the logs" });
+    }
   }
 });
 
 router.post("/edit/user/", isLoggedIn, async (req, res) => {
-  try {
-    let { _id, isAdmin, isSuperAdmin } = req.body;
-    let user = await User.findByIdAndUpdate(_id, {
-      $set: { isAdmin, isSuperAdmin },
-    });
-    res.status(200).json(user);
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ message: "something went south check the logs" });
+  if (!req.user.isSuperAdmin) {
+    res.status(403).json("Not Cool");
+  } else {
+    try {
+      let { _id, isAdmin, isSuperAdmin } = req.body;
+      let user = await User.findByIdAndUpdate(_id, {
+        $set: { isAdmin, isSuperAdmin },
+      });
+      res.status(200).json(user);
+    } catch (error) {
+      console.log(error);
+      res.status(500).json({ message: "something went south check the logs" });
+    }
+  }
+});
+
+router.post("/edit/project/", isLoggedIn, async (req, res) => {
+  if (!req.user.isSuperAdmin) {
+    res.status(403).json("Not Cool");
+  } else {
+    try {
+      let { _id, image, title, contributor, description, github } = req.body;
+      let project = await Project.findByIdAndUpdate(_id, {
+        $set: { image, title, contributor, description, github },
+      });
+      res.status(200).json(project);
+    } catch (error) {
+      console.log(error);
+      res.status(500).json({ message: "something went south check the logs" });
+    }
   }
 });
